Fix GenerateTownNameString signature
GenerateTownNameString had different signatures in the header file
and in the source file, and this caused trouble in 64-bit systems
where size_t is not unsigned int.

+++ src/townnamegen.cpp
@@ -1980,7 +1980,7 @@ assert_compile (lengthof(_town_name_generators) == N_ORIG_TOWN_NAME_GEN);
  * @param lang town name language
  * @param seed generation seed
  */
-void GenerateTownNameString (stringb *buf, size_t lang, uint32 seed)
+void GenerateTownNameString (stringb *buf, uint lang, uint32 seed)
 {
 	assert(lang < lengthof(_town_name_generators));
